# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EternaPy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found (needs onboarding)
    post:
      operationId: upsertUserProfile
      tags: [users]
      summary: Create or update user profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        subscriptionPlan:
          type: string
          enum: [free, pro, team]
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - username
        - displayName
        - xp
        - level
        - streak
        - subscriptionPlan
        - onboardingComplete
        - createdAt
    UserProfileInput:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
      required:
        - username
        - displayName
...